refactor: extract WorkspaceState group management into an extension file
Move the group lifecycle (details load/reload, profile + image editing,
invites, member admin mutations, archive/leave, group-image search, and the
group-member/details caches with their generation-token guards) into
WorkspaceState+Groups.swift. Pure code movement, no behavior change.
Build green; all 208 tests pass.
